+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+### Added
+- add option to check-3ware-status to specify the path to tw-cli binary
 
 ## [0.1.1] - 2016-09-01
 ### Added